Refactor nginx config into own package
